Refactor CCP dashboard component
